[US_AZ] Add metadata field to `state_sentence_group` (Recidiviz/recidiviz-data#50461)
## Description of the change

This adds a metadata field to the `state_sentence_group` entity to make
it easier to discern between different dates that we are squeezing into
the existing date fields. See
https://github.com/Recidiviz/recidiviz-data/issues/50459 and [this
(long)
thread](https://recidiviz.slack.com/archives/C063A244KC7/p1759265604840039)
for context.

This could be avoided if we could ingest probation release dates and
community supervision release dates into distinct fields in AZ, but
since we only have a handful of date fields and these dates are
(theoretically, and mostly realistically) mutually exclusive, we are
currently coalescing them into the existing fields. However, the dates
have different names; to display the correct copy in the frontend, we
need a way to tell if the release is to probation or community
supervision. That's what I'll be using this for!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6784bc5d11c896ca572bb976f6f1c03630e60884