[US_PA] Update IP and SP ingest for county jail periods (Recidiviz/recidiviz-data#50501)
## Description of the change

This PR aims to correct bugs in how we handle periods of time spent in a
county jail or on a type of warrant/detainer status in PA. Specifically,
it adds a branch to the if/then logic in the supervision type mappings
that checks for any detainer codes and assigns a value of
`WARRANT_STATUS` if found. This PR also adds inferred incarceration
periods in cases of supervision periods with a supervision level of
`IN_CUSTODY` using the default normalization logic.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#45335

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fe12504e1358e53a47103f3e9d49cfffc44968fd