Adds tasks contact type attribute to existing contact observation (Recidiviz/recidiviz-data#50549)
## Description of the change

Adds `tasks_contact_type` (representing the tool-specific contact type
label) to the `SUPERVISION_CONTACT` event observation so that we can
disaggregate `CONTACTS_ATTEMPTED` and `CONTACTS_COMPLETED` by the
contact type as its shown in the Tasks tool. Adds these two metrics to
the Tasks impact metrics Looker explore so that we can start tracking a
% missed contact rate around the launch of the JII appointment reminders
tool.

I added `tasks_contact_type` as the json field filter since this will
let us disaggregate and/or filter by the contact type within the
explore!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3dd0ce3756ad6da3fc36123a528336754ed31d20