[US_AZ] Fix bug in side panel mandatory literacy status query (Recidiviz/recidiviz-data#50723)
## Description of the change

We were filtering on an old version of `participation_status_raw_text`
that could be just the string `PARTICIPATING`. That is not a possible
raw text value for this field anymore, since we concatenated more values
to it in service of
https://github.com/Recidiviz/recidiviz-data/issues/47095 and
https://github.com/Recidiviz/recidiviz-data/issues/47634.

This makes the filter look for `PARTICIPATING` anywhere in the raw text,
instead of checking if the raw text is exactly equal to this string.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f604753033173949f30c6be1ed95f688944119d3