[US_NY] Update gender ingest (Recidiviz/recidiviz-data#50871)
## Description of the change

state_person gender was previously hydrated using the gendid field in
UNDER_CUSTODY. We recently started receiving monthly data (previously,
we just had a few yearly UNDER_CUSTODY files), and in these tables,
there are a number of possible values for gendid that weren't handled in
the mapping, causing ingest pipeline errors. These values aren't in the
codebook so we can't create mappings for them yet.

This change updates the state_person view to start using genid instead
of gendid to hydrate gender. This column has a more limited range of
possible values, only one of which ("0") is missing a mapping.

See followup ticket Recidiviz/recidiviz-data#50870 which was created because we need to find a
mapping for genid = "0", and potentially the numerous possible gendid
values if we end up using that column instead.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2b70afbc60baf60ce7a9cf9c96d30b9abb2799a1