[US_MO] Remap some supervision levels (Recidiviz/recidiviz-data#50876)
## Description of the change

We recently started seeing new supervision level codes, the most common
of which are LV1, LV2, and LV3 ("Level 1", "Level 2", and "Level 3").
These were mapped to INTERNAL_UNKNOWN since we haven't learned exactly
how to interpret these levels, but since they're so common and seem
likely to map to MINIMUM, MEDIUM, and HIGH, they've been updated to use
these mappings. We may need to revisit this later for Tasks, once we get
more updated policy information on contact standards for various
supervision levels.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e35edac275867e2b354f738d198549106dd31342