[US_AZ] Disambiguate release dates in resident record (Recidiviz/recidiviz-data#50904)
## Description of the change

This PR builds on https://github.com/Recidiviz/recidiviz-data/pull/50462
by pulling the `sentence_group_length_metadata` values into the results
of `us_az_projected_dates` and, further downstream,
`us_az_resident_metadata`.

I chose to do this by joining back to
`normalized_state.state_sentence_group_length` in the
`us_az_projected_dates` query because AZ is the only state that
currently needs or uses the metadata field as of now, so it doesn't feel
worth changing `sentence_sessions` views to include it.

In the course of implementing this, I found a couple of other bugs that
I also addressed here:
1) Partitioning by `sentence_group_external_id` instead of the
aggregated `sentence_inferred_group_id` was creating many overlapping
spans in the `us_az_projected_dates` view, which were propagating into
two TES queries and causing [a validation
failure](https://admin-panel-staging.recidiviz.org/admin/validation_metadata/status/details/overlapping_tes_criteria_spans?stateCode=US_AZ).
2) The `us_az_state_sentence_group` ingest mapping was setting default
values instead of leaving metadata fields NULL when neither of the
conditional date fields were hydrated. Now,
`parole_eligibility_date_external` in the metadata will be NULL if
neither of the CSBD or TR to ADD are set, and
`projected_full_term_release_date_min_external` will be NYULL if neither
ERCD or ADD are set.

One drawback of joining to metadata _after_ release dates have been
sessionized is that the release type cannot be taken into account during
sessionization. There are very few (<400) instances of a release date
staying the same, but the type of release changing from probation to
community supervision, or vice versa. Without accounting for the release
type when creating sessions, or recreating the sessions logic in this
view, we can't segment those changes.

The view currently handles those cases by choosing
`MAX(parole_eligibility_date_external)` and `MIN
(projected_full_term_release_date_min_external)` during aggregation.
This is a little confusing, but the idea is that there will only be
multiple non-null values across subsessions in this
multiple-release-type but same-release-date situation. When that
happens, this logic means we will display "Earned Release Credit Date"
and "Community Supervision Begin Date" on the frontend. I am comfortable
with that because it is the most common combination of dates, so it is
the most likely to be correct when we can't otherwise differentiate.

## Validation
AZ resident metadata has the same number of rows as it did before, and
fewer than 2% are different after this change. I spot-checked some of
the differences, and they don't seem more accurate. I made a ticket to
investigate further
https://github.com/Recidiviz/recidiviz-data/issues/50902. Since this
impacts a very small number of rows altogether, I will prioritize it
before the JII app launch, but I won't consider it blocking for this PR.

I loaded the ingest changes to a sandbox, and loaded the downstream
views off of that sandbox with prefix `elise_metadata3`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#50459

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1b8f0e93f12d27abe9dc8580a6a23022a2df3d02