[Sentences v2] Add test disallowing new references to state_incarceration_sentence_view, state_supervision_sentence_view, and state_charge_view (Recidiviz/recidiviz-data#50882)
## Description of the change

Updates the view tests to disallow new references to:
* `normalized_state.state_incarceration_sentence_view`
* `normalized_state.state_supervision_sentence_view`
* `normalized_state.state_charge_view`

As a part of this work, I updated a number of state-specific views to
pull from their corresponding state-specific `us_xx_normalized_state`
dataset view, rather than the state agnostic one. This will make it a
lot easier to track down which references we need to migrate as we work
to migrate each individual state.

All which should be deprecated as part of the v2 sentences migration.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#33402

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8f22620e2d3245413d1ea1791923475877252c67