[US_NY] Clean genid column in state_person view (Recidiviz/recidiviz-data#50895)
## Description of the change

For a handful of people, the `genid` column (used to hydrate
state_person gender) changes to 0 for a few UNDER_CUSTODY files. This
value doesn't have a definition in the codebook because it's only used
when there are data entry errors in the `gendid` column, which is used
to determine `genid`. We've confirmed that the correct approach in these
cases is just to use whatever `genid` value someone had before their
`genid` changed to 0: this change implements that step of cleaning.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 465f495ebdc419b86d1c068f80ccaf6ff44ead1c